backend/requirements.txt
charset-normalizer==3.3.2
click==8.1.7
Django==3.2.22
django-cors-headers==4.3.0
djangorestframework==3.14.0
djongo==1.3.6
fonttools==4.53.1
mongoengine==0.29.1
psycopg2==2.9.10
pycparser==2.22
Pygments==2.18.0
pymongo==3.12.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.32.3
sqlparse==0.2.4